pyproject.toml
[project]
name = "north-mcp-python-sdk"
version = "0.2.3"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Raphael Cristal", email = "raphael@cohere.com" }]
requires-python = ">=3.11"
dependencies = ["mcp>=1.9.0", "pyjwt[crypto]>=2.10.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "trio>=0.30.0",
]

[tool.uv]
constraint-dependencies = ["uv>=0.8.13"]

[tool.ruff]
line-length = 79
target-version = "py311"

[tool.ruff.format]
quote-style = "double"
indent-style = "space" 
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
